Bug 1408100 - Remove angular-ui-bootstrap from Perfherder Graphs dialogs
