+++ ui/entry-perf.js
@@ -58,6 +58,7 @@ require('./js/controllers/perf/alerts.js');
 require('./js/controllers/perf/dashboard.js');
 require('./js/controllers/perf/e10s-trend.js');
 require('./js/components/perf/compare.js');
+require('./js/components/perf/graphs.js');
 require('./js/components/auth.js');
 require('./js/components/loading.js');
 require('./js/perfapp.js');
