+++ ui/partials/perf/graphsctrl.html
@@ -12,17 +12,31 @@
               <span class="sr-only">Remove series</span>
             </button>
               <div class="graph-legend" style="border-left-color: {{series.blockColor}}" ng-class="{'series-inactive':!series.visible}">
-                <a href="" ng-click="addTestData('addRelatedConfigs', series.signature)" title="Add related configurations">{{series.name}}</a><br/>
-                <a href="" ng-click="addTestData('addRelatedBranches', series.signature)" title="Add related branches">{{series.projectName}}</a><br/>
-                <a href="" ng-click="addTestData('addRelatedPlatform', series.signature)" title="Add related platforms">{{series.platform}}</a><br/>
+                <a href="" data-target=#test-data-chooser-modal data-toggle="modal"
+                   data-option="addRelatedConfigs"
+                   data-series-signature="{{series.signature}}"
+                   title="Add related configurations">{{series.name}}</a><br/>
+                <a href="" data-target=#test-data-chooser-modal data-toggle="modal"
+                   data-option="addRelatedBranches"
+                   data-series-signature="{{series.signature}}"
+                   title="Add related branches">{{series.projectName}}</a><br/>
+                <a href="" data-target=#test-data-chooser-modal data-toggle="modal"
+                   data-option="addRelatedPlatform",
+                   data-series-signature="{{series.signature}}"
+                   title="Add related platforms">{{series.platform}}</a><br/>
                 <div class="signature"><small>{{series.signature}}</small></div>
               </div>
               <input title="Show/Hide series" type="checkbox" ng-model="series.visible" class="show-hide-check" ng-click="showHideSeries(series.signature)">
           </td>
         </tr>
       </table>
-      <button id="add-test-data-button" class="btn btn-primary" ng-click="addTestData()">
+      <button id="add-test-data-button"
+              class="btn btn-primary"
+              data-target=#test-data-chooser-modal
+              data-toggle="modal">
         <span class="glyphicon glyphicon-plus" aria-hidden="true"></span> Add <span ng-show="seriesList.length > 0">more</span> test data</button>
+      <test-data-chooser init-function="graphsDialogInit"
+                         success-callback="graphsDialogSuccessCallback"></test-data-chooser>
     </div>
     <div id="data-display">
       <select ng-model="myTimerange" ng-options="timerange.text for timerange in timeranges track by timerange.value" ng-change="timeRangeChanged()">
