--- ui/partials/perf/testdatachooser.html
@@ -1,67 +1,0 @@
-<div class="modal-header">
-  <button type="button" class="close" ng-click="cancel()"><span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button>
-  <h4 class="modal-title">Add test data</h4>
-</div>
-<div class="modal-body" ng-show="!loadingRelatedSignatures">
-    <p class="blink" >Getting related test information ...</p>
-</div>
-<div class="modal-body" ng-show="loadingRelatedSignatures">
-  <div id="performance-test-chooser" class="test-chooser">
-    <div class="form-group">
-      <label>Framework</label>
-      <select class="form-control" ng-change="updateTestInput()" ng-model="selectedFramework" ng-options="framework.name for framework in frameworkList">
-      </select>
-    </div>
-    <div class="form-group">
-      <label>Project</label>
-      <select class="form-control" ng-change="updateTestInput()" ng-model="selectedProject" ng-options="project.name for project in projects">
-      </select>
-    </div>
-    <div class="form-group">
-      <label>Platform</label>
-      <p class="blink" ng-show="loadingPlatformList">Loading platform list...</p>
-      <select class="form-control" ng-hide="loadingPlatformList" ng-change="updateTestSelector()" ng-model="selectedPlatform" ng-options="platform for platform in platformList">
-      </select>
-    </div>
-    <div class="form-group">
-      <label>Tests</label>
-      <div class="test-loading-placeholder" ng-show="loadingTestData">
-        <p class="blink">Loading series data...</p>
-      </div>
-      <input class="form-control input-sm" type="text" ng-model="testFilter" placeholder="Filter tests" ng-model-options="{debounce: 250}" ng-hide="loadingTestData"/>
-      <select ng-hide="loadingPlatformList || loadingTestData" multiple class="form-control choose-test-list" ng-model="selectedTestSignatures" ng-hide="loadingTestData">
-        <option value="{{::testElem.signature}}" ng-repeat="testElem in unselectedTestList| testNameContainsWords: testFilter track by testElem.signature" title="{{::testElem.name}}">
-          {{::testElem.name}}
-        </option>
-      </select>
-    </div>
-    <div class="checkbox">
-      <label>
-        <input type="checkbox" ng-model="includeSubtests" ng-change="updateTestSelector()">Include subtests</input>
-      </label>
-    </div>
-  </div>
-  <div class="btn-group-vertical">
-    <button id="unselect-test" ng-click="unselectTest()" type="button" class="btn btn-xs btn-default" ng-disabled="!testsToAdd.length">
-      <span class="glyphicon glyphicon-chevron-left"></span>
-    </button>
-    <button id="select-test" ng-click="selectTest()" type="button" class="btn btn-xs btn-default" ng-disabled="!unselectedTestList.length">
-      <span class="glyphicon glyphicon-chevron-right"></span>
-    </button>
-  </div>
-  <div class="test-list-container">
-    <div id="added-test-list" class="form-group">
-      <label>Tests to add</label>
-      <select multiple class="form-control" ng-model="selectedTestsToAdd">
-        <option value="{{testToAdd}}" ng-repeat="testToAdd in testsToAdd"
-                title="{{testToAdd.projectName}} {{testToAdd.platform}} {{testToAdd.name}}">
-           {{testToAdd.projectName}} {{testToAdd.platform}}
-           {{testToAdd.name}}
-        </option>
-      </select>
-    </div>
-  </div>
-</div>
-<div class="modal-footer">
-  <button class="btn btn-primary" ng-click="addTestData()" ng-disabled="!testsToAdd.length" ng-model="rightList"><span class="glyphicon glyphicon-plus" aria-hidden="true"></span> Add</button>
-</div>
